add dataSection and logicSection parsing
The commit uses regex to get the .DATA and .LOGIC section. However, there is currently a problem for parsing multi-word commands (e.g. SCAN RIGHT). Moreover, instead of just storing the data type, we want to make sure it stores the actual initialized data. It should be found in the other branch.